src/main/java/com/nhnacademy/shoppingmall/category/service/impl/CategoryServiceImpl.java
package com.nhnacademy.shoppingmall.category.service.impl;

import com.nhnacademy.shoppingmall.category.domain.Category;
import com.nhnacademy.shoppingmall.category.repository.CategoryRepository;
import com.nhnacademy.shoppingmall.category.service.CategoryService;

import java.util.List;
import java.util.Optional;

public class CategoryServiceImpl implements CategoryService {
    private final CategoryRepository categoryRepository;

    public CategoryServiceImpl(CategoryRepository categoryRepository) {
        this.categoryRepository = categoryRepository;
    }

    @Override
    public Optional<Category> getCategoryById(int categoryId) {
        return categoryRepository.findById(categoryId);
    }

    @Override
    public List<Category> getAllCategories() {
        return categoryRepository.findAll();
    }

    @Override
    public int addCategory(Category category) {

        return categoryRepository.save(category);
    }

    @Override
    public int updateCategory(Category category) {
        return categoryRepository.update(category);
    }

    @Override
    public int deleteCategory(int categoryId) {
        return categoryRepository.deleteById(categoryId);
    }
}
